Fixed PWM timing

+++ Drivers/Pwm/pwm/h/pwm.h
@@ -9,7 +9,7 @@ class PWM {
 private:
 
 public:
-    void cycle(uint8_t timer, uint16_t duty_cicle);
+    void cycle(uint8_t timer, uint32_t duty_cicle);
     void initServo(void);
     void initMotor(void);
 };
